+++ src/Network/Network/FirewallPolicy/PolicySettings/AzureApplicationGatewayFirewallPolicySetting.cs
@@ -68,6 +68,10 @@ namespace Microsoft.Azure.Commands.Network
         [ValidateNotNullOrEmpty]
         public string CustomBlockResponseBody { get; set; }
 
+        [Parameter(Mandatory = false, HelpMessage = "To scrub sensitive log fields")]
+        [ValidateNotNullOrEmpty]
+        public PSApplicationGatewayFirewallPolicyLogScrubbingConfiguration LogScrubbing { get; set; }
+
         public override void ExecuteCmdlet()
         {
             base.ExecuteCmdlet();
@@ -132,6 +136,7 @@ namespace Microsoft.Azure.Commands.Network
                 FileUploadLimitInMb = this.MaxFileUploadInMb,
                 CustomBlockResponseBody = this.CustomBlockResponseBody,
                 CustomBlockResponseStatusCode = this.CustomBlockResponseStatusCode,
+                LogScrubbing = this.LogScrubbing
             };
         }
     }